requirements/prod.txt
# Everything needed in production
djangorestframework==3.10.3
drf-yasg==1.17.0
django-filter==2.2.0
djangorestframework-simplejwt==4.3.0